added to predocx variables to stop 504 error
i tested the interview on my server and kept getting 504 errors when i tried to produce the document.  I then tried it on the lemma server, and saw that it took 25 seconds to complete processing, and a lot of time was being used to define variables after being called in the document.  i moved each variable that was being called to the predocx variables code block.  That cut the time on the lemma server to 17 seconds, and i could run the interview on my server without a 504 error, though it still took over 40 seconds so probably still not reliable.